refactor: change lib/api folder to lib/action
to avoid confusion between api backend and fetch api
