+++ lib/action/admin/patch-teacher/index.ts
